microservice/requirements.txt
Flask
pymongo
numpy
scipy
python-dotenv